+++ server/src/java/main/com/amazon/pipeitor/RadioListener.java
@@ -4,5 +4,5 @@ import com.rapplogic.xbee.api.zigbee.ZNetTxStatusResponse;
 
 public interface RadioListener {
     void handleDataPacket(Radio radio, RemoteAddress remoteAddress, byte[] data);
-    void handleTxStatusPacket(XBeeRadio radio, ZNetTxStatusResponse.DeliveryStatus status, int frameId);
+    void handleTxStatusPacket(XBeeRadio radio, ZNetTxStatusResponse.DeliveryStatus status, byte frameId);
 }
